Stylistic changes.
- Added an include to be explicit instead of implicit.
- Added static_cast<size_t> around std::streamsize for memcpy
  invocation.

+++ src/reader.cpp
@@ -11,6 +11,7 @@
 #include <cstring>
 #include <fstream>
 #include <condition_variable>
+#include <atomic>
 
 #include "bitdiff/reader.hpp"
 
@@ -120,7 +121,7 @@ size_t bd::Reader::read(unsigned char * buffer)
     std::streamsize ret = 0;
     if (m_read > 0)
     {
-        memcpy(buffer, m_buffer, m_read);
+        memcpy(buffer, m_buffer, static_cast<size_t>(m_read));
         ret = m_read;
         m_read = 0;
     }
